Refactor /webapp per agreed structure
