+++ tensorboard/components/BUILD
@@ -71,26 +71,6 @@ tensorboard_html_binary(
 )
 
 tf_web_library(
-    name = "ng_tensorboard",
-    srcs = [
-        "ng_tensorboard.html",
-    ],
-    path = "/",
-    deps = [
-        ":ng_polymer_lib_binary",
-        "//tensorboard/webapp",
-        "//tensorboard/webapp:material_theme",
-    ],
-)
-
-tensorboard_html_binary(
-    name = "ng_index",
-    input_path = "/ng_tensorboard.html",
-    output_path = "/ng_index.html",
-    deps = [":ng_tensorboard"],
-)
-
-tf_web_library(
     name = "trace_viewer",
     srcs = ["trace_viewer.html"],
     path = "/",
