+++ tensorboard/webapp/core/BUILD
@@ -7,15 +7,20 @@ load("//tensorboard/defs:defs.bzl", "tf_ts_library")
 ng_module(
     name = "core",
     srcs = [
-        "core.actions.ts",
-        "core.effects.ts",
+        "actions/core.actions.ts",
+        "actions/index.ts",
+        "effects/core.effects.ts",
+        "effects/index.ts",
         "core.module.ts",
-        "core.reducers.ts",
-        "core.service.ts",
+        "store/core.reducers.ts",
+        "store/core.selectors.ts",
+        "store/core.types.ts",
+        "store/index.ts",
     ],
     deps = [
         "//tensorboard/webapp/angular:expect_angular_common_http",
         "//tensorboard/webapp/types",
+        "//tensorboard/webapp/core_service",
         "@npm//@angular/common",
         "@npm//@ngrx/effects",
         "@npm//@ngrx/store",
@@ -27,13 +32,14 @@ tf_ts_library(
     name = "core_effects_test_lib",
     testonly = True,
     srcs = [
-        "core.effects.test.ts",
+        "effects/core.effects.test.ts",
     ],
     tsconfig = "//:tsconfig-test",
     deps = [
         ":core",
         "//tensorboard/webapp/angular:expect_angular_common_http_testing",
         "//tensorboard/webapp/angular:expect_angular_core_testing",
+        "//tensorboard/webapp/core_service",
         "//tensorboard/webapp/core/testing",
         "//tensorboard/webapp/types",
         "@npm//@angular/common",
@@ -50,7 +56,7 @@ ng_module(
     name = "core_reducers_test_lib",
     testonly = True,
     srcs = [
-        "core.reducers.test.ts",
+        "store/core.reducers.test.ts",
     ],
     deps = [
         ":core",
