+++ tensorboard/webapp/core/store/core.reducers.test.ts
@@ -15,10 +15,10 @@ limitations under the License.
 import {expect} from 'chai';
 import * as sinon from 'sinon';
 
-import * as actions from './core.actions';
+import * as actions from '../actions';
 import {reducers} from './core.reducers';
-import {createPluginMetadata, createCoreState} from './testing';
-import {LoadState} from '../types/api';
+import {createPluginMetadata, createCoreState} from '../testing';
+import {LoadState} from '../../types/api';
 
 function createPluginsListing() {
   return {
