+++ tensorboard/webapp/core/store/index.ts
@@ -12,21 +12,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-import {Component, OnInit} from '@angular/core';
-import {Store} from '@ngrx/store';
-import {coreLoaded} from './core/actions';
 
-/** @typehack */ import * as _typeHackRxjs from 'rxjs';
-
-@Component({
-  selector: 'tb-webapp',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
-})
-export class AppComponent implements OnInit {
-  constructor(private readonly store: Store<{}>) {}
-
-  ngOnInit() {
-    this.store.dispatch(coreLoaded());
-  }
-}
+export * from './core.reducers';
+export * from './core.selectors';
