+++ tensorboard/webapp/header/header.container.ts
@@ -15,16 +15,13 @@ limitations under the License.
 import {Component} from '@angular/core';
 import {Store, select, createSelector} from '@ngrx/store';
 
-import {State, getActivePlugin, getPlugins} from '../../core/core.reducers';
-import {changePlugin} from '../../core/core.actions';
+import {getActivePlugin, getPlugins} from '../core/store';
+import {changePlugin} from '../core/actions';
 
-import {PluginMetadata, PluginId} from '../../types/api';
+import {PluginId} from '../types/api';
+import {UiPluginMetadata} from './types';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
-
-export interface UiPluginMetadata extends PluginMetadata {
-  id: PluginId;
-}
 
 const getUiPlugins = createSelector(
   getPlugins,
@@ -55,7 +52,7 @@ export class HeaderContainer {
   readonly plugins$ = this.store.pipe(select(getUiPlugins));
   readonly disabledPlugins$ = this.store.pipe(select(getDisabledPlugins));
 
-  constructor(private readonly store: Store<State>) {}
+  constructor(private readonly store: Store<{}>) {}
 
   onPluginSelectionChange(pluginId: PluginId) {
     this.store.dispatch(changePlugin({plugin: pluginId}));
