+++ tensorboard/webapp/header/header.module.ts
@@ -20,8 +20,8 @@ import {MatTabsModule} from '@angular/material/tabs';
 import {MatToolbarModule} from '@angular/material/toolbar';
 import {MatSelectModule} from '@angular/material/select';
 
-import {HeaderComponent} from './components/header.component';
-import {HeaderContainer} from './containers/header.component';
+import {HeaderComponent} from './header.component';
+import {HeaderContainer} from './header.container';
 import {CoreModule} from '../core/core.module';
 import {SettingsModule} from '../settings/settings.module';
