+++ tensorboard/webapp/header/header.test.ts
@@ -22,11 +22,10 @@ import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 import {Store} from '@ngrx/store';
 import {provideMockStore, MockStore} from '@ngrx/store/testing';
 
-import {HeaderComponent} from './components/header.component';
-import {HeaderContainer} from './containers/header.component';
+import {HeaderComponent} from './header.component';
+import {HeaderContainer} from './header.container';
 
-import {changePlugin} from '../core/core.actions';
-import {State} from '../core/core.reducers';
+import {changePlugin} from '../core/actions';
 import {
   createPluginMetadata,
   createState,
@@ -37,7 +36,7 @@ import {PluginId} from '../types/api';
 /** @typehack */ import * as _typeHackStore from '@ngrx/store';
 
 describe('header test', () => {
-  let store: MockStore<State>;
+  let store: MockStore<{}>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
