+++ tensorboard/webapp/index.html
@@ -30,9 +30,6 @@ limitations under the License.
 <link rel="stylesheet" href="@angular/indigo-pink.css" />
 
 <body>
-  <script
-    jscomp-nocompile
-    src="tf-ng-tensorboard/tf_ng_tensorboard_binary.js"
-  ></script>
-  <tf-ng-tensorboard></tf-ng-tensorboard>
+  <script jscomp-nocompile src="tb-webapp/tb_webapp_binary.js"></script>
+  <tb-webapp></tb-webapp>
 </body>
